requirements.txt
babel==2.17.0
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
python-dateutil==2.9.0.post0
six==1.17.0
tkcalendar==1.6.1
